Nudged around packet readers and writers, added placeholders.
Removed `EzSocketsPacketReader`; I've determined that it probably misses the point of this driver, which is to offload the difficult parts of the driver, of which network connectivity is probably one, to another program. `InputHandler_SextetStream` support for the driver has been removed accordingly.

`SelectFilePacketReader` becomes `PosixSelectFifoPacketReader` to emphasize the target platform.

A placeholder is added for `PosixSelectFifoPacketWriter`, which is to be implemented in much the same idiom as `SelectFilePacketReader`.

Placeholders are added for `WindowsOverlappedPipePacketReader` and `WindowsOverlappedPipePacketWriter`. These will operate in the same fashion as `PosixSelectFifoPacket*`, but use the Windows API facilities `OVERLAPPED` and `WaitForSingleObject()` which provide a close analogue to `select()` with timeout.

`RageFilePacketWriter` is removed, and a placeholder is added for `StdCFilePacketWriter`, which will serve as the new fallback for systems that don't claim to be Windows or POSIXish. (Not sure which systems these are.)